add shutdown_intf command for TopologyOpsCommander
